Flex body and width contents
After the making of the skeleton of the landing page I give all a border and flex the body to create a width that fit the max width. Next is that I create another divs that wrap the contents so I can centered and match all of the individual divs. With each content wrap around a width-content class it won't interfere in creating the background colors of each contents.